+++ lib/app/core/models/series_detail.dart
@@ -1,3 +1,4 @@
+
 import 'package:cm_app/app/core/models/movie_cast.dart';
 import 'package:cm_app/app/core/models/season.dart';
 import 'package:than_pkg/services/index.dart';